requirements.txt
numpy>=1.26,<2.0
pywebview>=4.4
fastapi>=0.111
uvicorn>=0.30
python-multipart>=0.0.9
httpx>=0.27
pymupdf>=1.24
python-docx>=1.1
ultralytics>=8.2
easyocr>=1.7
opencv-python>=4.9